requirements.txt
Django==5.0.3
psycopg2-binary==2.9.9
Pillow==10.2.0
python-dotenv==1.0.1